requirements.txt
pandas==2.0.1
argparse==1.1
dnspython==2.7.0
numpy==1.26.4
git+https://github.com/jamesdolan/ip_to_country.git@main
